Make `IdentityTransformation3D` an outer class
